Update power suppy info action
This action now checks the status of the power supply before responding.
If the power supply is down, the status of the occurrence is described.

If it is up, the action marks that power supply as "pending", awaiting
another user report.

If it is pending, the power supply is assumed to be down and the
resolution process is started (though it is just a mock at this stage)